requirements.txt
streamlit>=1.20
pymupdf>=1.22
faiss-cpu>=1.7.3
numpy>=1.24
cohere>=3.0.0
requests>=2.28
python-dotenv>=1.0
pdfplumber>=0.7
pytesseract>=0.3
Pillow>=9.0
